tui: colocate app module tests
